First draft: fixing the problems with relative and absolute file paths

+++ test/LargeTests/test_large.py
@@ -6,33 +6,75 @@ from TimeLoopLib import TreeDynamicTimeStepping
 import unittest
 import glob
 import os
+from lxml import etree
+import shutil
 
-filepath_example_setups = path.join(path.dirname(path.abspath(__file__)),"Test_Setups_large/*.xml")
-xml = glob.glob(filepath_example_setups)
+filepath_examplesetups = path.join(path.dirname(path.abspath(__file__)),"Test_Setups_large/*.xml")
+xml = glob.glob(filepath_examplesetups)
 errors = []
+global output_exist
+output_exist = str
 if xml:
-    for xml_file in xml:
-        filepath_results = path.join(path.dirname(path.abspath(__file__)),"Test_Setups_large/testoutputs/*.*")
-        results = glob.glob(filepath_results)
-        for result in results:
-            os.remove(result)
+    for xmlfile in xml:
         print("________________________________________________")
-        print("In the following the setup", xml_file, "is tested.")
-    
+        print("In the following the setup", xmlfile, "is tested.")
+        print("________________________________________________")
+
+
+        def findChild( parent, key):
+            child = parent.find(key)
+            return child
+        
+        global output_dir
+        global output_type
+        
+        tree = etree.parse(xmlfile)
+        root = tree.getroot()
+        for tag in root.iter():
+            tag.text = tag.text.strip()
+            
+        output = findChild(root, "tree_output")
+        output_type_xml_element = findChild(output, "type")
+        output_type = output_type_xml_element.text
+        
+        if not output_type == "NONE":
+            output_dir_xml_element = findChild(output, "output_dir")
+            #output_dir = path.join(path.dirname(path.abspath(__file__)),output_dir_xml_element.text)
+            output_dir = output_dir_xml_element.text
+            
+            if not os.path.exists(output_dir):
+                output_exist = "n"                    
+                os.makedirs(output_dir)
+            else:
+                output_exist = "y"
+                old_results = glob.glob(path.join(output_dir,"*.*"))
+                if old_results:
+                    for result in old_results:
+                        os.remove(result)
+        else:
+                output_exist = "e"
+                
         class MyTest(unittest.TestCase):
-    
             def test(self):
                 try:
-                    prj = XMLtoProject(xml_project_file=xml_file)
+                    prj = XMLtoProject(xml_project_file=xmlfile)
                     time_stepper = TreeDynamicTimeStepping(prj)
                     prj.runProject(time_stepper)
                 except:
-                    self.fail(errors.append(xml_file))
-    
+                    self.fail(errors.append(xmlfile))
+        
         if __name__ == "__main__":
             unittest.main(exit=False)
-        print("The setup", xml_file, "was tested.")
-        print("________________________________________________")
+        if not output_type == "NONE":
+            if output_exist == "n":
+                shutil.rmtree((output_dir[:-1]), ignore_errors=True)
+            if output_exist == "y":
+                old_results = glob.glob(path.join(output_dir,"*.*"))
+                for result in old_results:
+                    os.remove(result)
+        
+    print("The setup", xmlfile, "was tested.")
+    print("________________________________________________")
     print("The testing of all example setups is finished.")
     print("")
     if errors:
@@ -48,4 +90,4 @@ if xml:
     else:
         print("The tests of all example setups were successful.")
         print("")
-else: print("unfortunately no project-file could be found.")
+else: print("Unfortunately no project-file could be found.")
